Alterações no front e correção de bugs

+++ backend/vendedor/salvarController.php
@@ -6,7 +6,7 @@ use App\Model\Vendedor;
 $vendedor = new Vendedor();
 // var_dump($_POST);die;
 $padrao = '/[^0-9]/i';
-$_POST['cpf'] = preg_replace($padrao, '', $_POST['cpf']);
+$_POST['cpf'] = (int)preg_replace($padrao, '', $_POST['cpf']);
 
 // var_dump($_POST['cpf']);die;
 $vendedor->salvar($_POST);
